+++ src/main/resources/static/js/staff/staffSuggest.js
@@ -12,4 +12,13 @@ $(document).ready(function(){
             $('#staffSearchForm').submit();
         }})
         .select();
+
+    $("#voucherStaffSearch")
+        .autocomplete({
+            serviceUrl: '/voucher/suggest',
+        onSelect: function (suggestion) {
+            $('#voucherStaffSearch').val(suggestion.value);
+            $('#voucherStaffSearchForm').submit();
+        }})
+        .select();
 });